@@ -1,25 +1,25 @@
 name: golden_toolkit
 description: Common patterns for screenshot-based widget testing using Goldens.
-version: 0.15.0
+version: 0.16.0
 homepage: https://github.com/eBay/flutter_glove_box/
 repository: https://github.com/eBay/flutter_glove_box/tree/master/packages/golden_toolkit
 issue_tracker: https://github.com/eBay/flutter_glove_box/issues
 
 environment:
-  sdk: ">=2.18.4 <3.0.0"
+  sdk: ^3.9.2
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  meta: ^1.8.0
+  meta: ^1.16.0
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: ^6.0.0
   sample_dependency:
     path: test/sample_dependency
-  test_api: ^0.4.12
+  test_api: ^0.7.6
 
 flutter:
   fonts:
